Move Trip and Stop entities to backend-schedule module
