+++ backend/backend-schedule/src/main/java/eu/socialedge/hermes/backend/schedule/domain/gen/basic/BasicScheduleGenerator.java
@@ -20,8 +20,8 @@ import eu.socialedge.hermes.backend.schedule.domain.gen.ScheduleGenerator;
 import eu.socialedge.hermes.backend.schedule.domain.gen.ScheduleGeneratorException;
 import eu.socialedge.hermes.backend.transit.domain.Line;
 import eu.socialedge.hermes.backend.transit.domain.Route;
-import eu.socialedge.hermes.backend.transit.domain.Stop;
-import eu.socialedge.hermes.backend.transit.domain.Trip;
+import eu.socialedge.hermes.backend.schedule.domain.Stop;
+import eu.socialedge.hermes.backend.schedule.domain.Trip;
 import lombok.*;
 import tec.uom.se.unit.Units;
